Test code tidy up
